+++ README.md
@@ -11,6 +11,6 @@
 
 - [x] 게임의 참여할 사람을 입력받는다.
 - [x] 딜러와 플레이어들이 카드를 나눠 받는다.
-- [ ] 딜러와 플레이어들이 추가 카드를 나눠 갖는다.
-- [ ] 딜러와 플레이어들의 결과를 출력한다.
-- [ ] 최종 결과를 출력한다.
+- [x] 딜러와 플레이어들이 추가 카드를 나눠 갖는다.
+- [x] 딜러와 플레이어들의 결과를 출력한다.
+- [x] 최종 결과를 출력한다.
