+++ src/main/java/domain/game/BlackjackSimulator.java
@@ -1,6 +1,9 @@
-package domain;
+package domain.game;
 
+import domain.handouter.*;
 import domain.validator.ObjectsValidator;
+import domain.winorlose.BlackjackWinOrLoseDecider;
+import domain.winorlose.FinalWinOrLose;
 import view.input.InputView;
 import view.output.OutputView;
 import view.output.dto.BlackjackResultOutputs;
