+++ src/main/java/domain/participant/BlackjackDealer.java
@@ -1,6 +1,9 @@
-package domain;
+package domain.participant;
 
+import domain.trumpcard.TrumpCard;
+import domain.trumpcard.TrumpCards;
 import domain.validator.ObjectsValidator;
+import domain.winorlose.BlackjackPoint;
 
 import java.util.List;
