+++ src/main/java/domain/participant/BlackjackParticipants.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 import util.CollectionUtils;
