+++ src/main/java/domain/participant/BlackjackPlayerName.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 import util.StringUtils;
