+++ src/main/java/domain/trumpcard/TrumpCardSuit.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.trumpcard;
 
 public enum TrumpCardSuit {
