+++ src/main/java/domain/winorlose/PlayerWinOrLose.java
@@ -1,4 +1,7 @@
-package domain;
+package domain.winorlose;
+
+import domain.participant.BlackjackPlayerName;
+import domain.trumpcard.WinOrLose;
 
 import java.util.Objects;
