+++ src/main/java/view/input/dto/HandOutPlayerRequest.java
@@ -1,6 +1,6 @@
 package view.input.dto;
 
-import domain.BlackjackPlayerName;
+import domain.participant.BlackjackPlayerName;
 
 public class HandOutPlayerRequest {
