+++ src/main/java/view/output/dto/InitialHandOutParticipantOutput.java
@@ -1,8 +1,8 @@
 package view.output.dto;
 
-import domain.BlackjackDealer;
-import domain.BlackjackParticipant;
-import domain.BlackjackPlayer;
+import domain.participant.BlackjackDealer;
+import domain.participant.BlackjackParticipant;
+import domain.participant.BlackjackPlayer;
 
 import java.util.List;
